[REFACT]: FeedHistoryAPI -> FeedAPI로 통합
