+++ Projects/App/Sources/AppContainer.swift
@@ -100,10 +100,6 @@ final class AppContainer:
     return EndedChallengeUseCaseImpl(repository: challengeRepository)
   }()
   
-  lazy var feedHistoryUseCase: FeedHistoryUseCase = {
-    return FeedHistoryUseCaseImpl(repository: feedHistoryRepository)
-  }()
-  
   lazy var homeUseCae: HomeUseCase = {
     return HomeUseCaseImpl(repository: challengeRepository)
   }()
@@ -183,9 +179,5 @@ final class AppContainer:
 
   lazy var resignRepository: ResignRepository = {
     return ResignRepositoryImpl(dataMapper: ResignDataMapperImpl())
-  }()
-  
-  lazy var feedHistoryRepository: FeedHistoryRepository = {
-    return FeedHistoryRepositoryImpl(dataMapper: FeedHistoryDataMapperImpl())
   }()
 }
