+++ Projects/Data/DataMapper/ChallengeDataMapper.swift
@@ -21,6 +21,7 @@ public protocol ChallengeDataMapper {
   func mapToFeedComment(dto: CommentResponseDTO) -> FeedComment
   func mapToChallengeDescription(dto: ChallengeDescriptionResponseDTO, id: Int) -> ChallengeDescription
   func mapToChallengeMembers(dto: [ChallengeMemberResponseDTO]) -> [ChallengeMember]
+  func mapToFeedHistory(dto: FeedHistoryResponseDTO) -> [FeedHistory]
 }
 
 public struct ChallengeDataMapperImpl: ChallengeDataMapper {
@@ -158,6 +159,18 @@ public struct ChallengeDataMapperImpl: ChallengeDataMapper {
         isOwner: $0.isCreator,
         duration: $0.duration,
         goal: $0.goal
+      )
+    }
+  }
+  
+  public func mapToFeedHistory(dto: FeedHistoryResponseDTO) -> [FeedHistory] {
+    return dto.content.map {
+      FeedHistory(
+        feedId: $0.feedId,
+        challengeId: $0.challengeId,
+        imageUrl: $0.imageUrl,
+        provedDate: $0.createDate,
+        name: $0.name
       )
     }
   }
