--- Projects/Domain/UseCase/Implementations/FeedHistoryUseCaseImpl.swift
@@ -1,24 +1,0 @@
-//
-//  FeedHistoryUseCaseImpl.swift
-//  Domain
-//
-//  Created by 임우섭 on 2/23/25.
-//  Copyright © 2025 com.photi. All rights reserved.
-//
-
-import RxSwift
-import Entity
-import UseCase
-import Repository
-
-public struct FeedHistoryUseCaseImpl: FeedHistoryUseCase {
-  private let repository: FeedHistoryRepository
-  
-  public init(repository: FeedHistoryRepository) {
-    self.repository = repository
-  }
-  
-  public func fetchFeeds(page: Int, size: Int) -> Single<[FeedHistory]> {
-    return repository.fetchFeedHistory(page: page, size: size)
-  }
-}
