--- Projects/Domain/UseCase/Interfaces/FeedHistoryUseCase.swift
@@ -1,14 +1,0 @@
-//
-//  FeedHistoryUseCase.swift
-//  Domain
-//
-//  Created by 임우섭 on 2/23/25.
-//  Copyright © 2025 com.photi. All rights reserved.
-//
-
-import RxSwift
-import Entity
-
-public protocol FeedHistoryUseCase {
-  func fetchFeeds(page: Int, size: Int) -> Single<[FeedHistory]>
-}
