+++ Projects/Domain/UseCase/Interfaces/FeedUseCase.swift
@@ -21,4 +21,5 @@ public protocol FeedUseCase {
   func uploadFeedComment(challengeId: Int, feedId: Int, comment: String) async throws -> Int
   func deleteFeedComment(challengeId: Int, feedId: Int, commentId: Int) async throws
   func deleteFeed(challengeId: Int, feedId: Int) -> Single<Void>
+  func fetchFeeds(page: Int, size: Int) -> Single<[FeedHistory]>
 }
